+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/reroute/NavigationRerouteController.kt
@@ -1,5 +1,6 @@
 package com.mapbox.navigation.core.reroute
 
+import androidx.annotation.UiThread
 import com.mapbox.navigation.base.route.NavigationRoute
 import com.mapbox.navigation.base.route.RouterOrigin
 import com.mapbox.navigation.core.MapboxNavigation
@@ -9,6 +10,7 @@ import com.mapbox.navigation.core.trip.session.OffRouteObserver
  * Reroute controller allows changing the reroute logic externally. Use [MapboxNavigation.rerouteController]
  * to replace it.
  */
+@UiThread
 interface NavigationRerouteController : RerouteController {
 
     /**
@@ -23,6 +25,7 @@ interface NavigationRerouteController : RerouteController {
      * Route Callback is useful to set new route(s) on reroute event. Doing the same as
      * [MapboxNavigation.setRoutes].
      */
+    @UiThread
     fun interface RoutesCallback {
         /**
          * Called whenever new route(s) has been fetched.